requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow==10.0.1
psutil
py-tgcalls==0.9.7
pykeyboard
pyrofork
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp
youtube-search-python
py_yt_search
pymongo==4.6.3
httpx[http2]>=0.28.1
py-yt
